requirements.txt
Flask==2.2.3
Pillow==10.0.0
google-generativeai==0.7.2
werkzeug==2.3.0
